requirements.txt
psycopg2-binary
pymysql
pymongo
redis